Add `delta` into steering velocity calculations
I'm not completely convinced that this is required at this time. Just
from some basic scratch math on the side, the units don't seem quite
right to me, but it does change the maximum value we would apply to
`acceleration_coefficient` and we just compensate there.

It's not causing harm for now, so until I get confirmation that the
units are all right regardless of frame rate, we'll put it in.`